Issue ekstep/Learning-Platform-Analytics#1144 feat: implement cloud-storage-sdk

+++ src/main/scala/org/sunbird/cloud/storage/Model.scala
@@ -4,5 +4,5 @@ import java.util.Date
 
 object Model {
   
-    case class Blob(key: String, contentLength: Int, lastModified: Date, metadata: Map[String, AnyRef], payload: Option[Array[Byte]]);
+    case class Blob(key: String, contentLength: Long, lastModified: Date, metadata: Map[String, AnyRef], payload: Option[Array[Byte]]);
 }
